fix(substreams): split eth_methods into one wasm binary per RPC extension
A wasm module resolves its imports when it is instantiated rather than
when it calls them, so the single binary the package used to ship was
refused wholesale by any instance registering only one of the two RPC
extensions:

    module "map_transfers": general wasm execution failed: could not
    instantiate wasm module: unknown import: `rpc::eth_get_balance`
    has not been defined

That failure hit every module of the package, including `map_transfers`,
which performs no RPC at all and is documented as the one to run first to
tell a decoding problem apart from an RPC one.

`fireeth` registers each extension independently, `rpc::eth_call` on
`--substreams-rpc-endpoints` and `rpc::eth_get_balance` on
`--substreams-rpc-get-balance-endpoints`, so the handlers now live in one
binary per extension plus a third importing neither:

    no_rpc.wasm           map_transfers, map_eth_methods    no extension
    eth_call.wasm         map_eth_call                      rpc::eth_call
    eth_get_balance.wasm  map_eth_get_balance               rpc::eth_get_balance

`substreams.yaml` gained the matching `binaries:` entries and points each
module at the one it belongs to. A missing endpoint flag now fails the
module that needs it and nothing else. `map_eth_methods` stays the
exception, not through its own binary but because it consumes the output
of both RPC modules.

Handlers moved to `src/bin`, one file per binary, and `src/lib.rs` keeps
only what they share. Cargo builds every `src/bin` target on a plain
`cargo build`, which is what `substreams build` runs, so the three wasm
files still come out of a single invocation and the Makefile is unchanged.
