+++ substreams/modules/eth_methods/src/lib.rs
@@ -1,19 +1,23 @@
-//! Diagnostic Substreams exercising the two RPC WASM extensions a `fireeth` instance registers,
-//! `rpc::eth_call` and `rpc::eth_get_balance`.
+//! Code shared by the binaries under `src/bin`, which is where the handlers live.
 //!
-//! Each extension is enabled by its own `fireeth` flag, `--substreams-rpc-endpoints` and
-//! `--substreams-rpc-get-balance-endpoints`, and an operator may well have configured only one
-//! of them. They are therefore exercised by two separate modules so that a missing endpoint
-//! fails the module that needs it instead of the whole package. `map_eth_methods` runs both for
-//! the common case where everything is expected to work.
+//! This package exercises the two RPC WASM extensions a `fireeth` instance registers,
+//! `rpc::eth_call` and `rpc::eth_get_balance`. Each is enabled by its own flag,
+//! `--substreams-rpc-endpoints` and `--substreams-rpc-get-balance-endpoints`, and an operator may
+//! well have configured only one of them, or neither.
+//!
+//! A WASM module resolves its imports when it is instantiated rather than when it calls them, so a
+//! single binary referencing both extensions fails to instantiate whenever one of the two is
+//! unregistered, and it fails for every module of the package, including the ones performing no
+//! RPC at all. Splitting the handlers over one binary per extension is what keeps a missing
+//! endpoint failing the modules that need it instead of all of them.
+//!
+//! Nothing here may reference either extension, directly or transitively: this library is linked
+//! into all three binaries, and a reference would put the import back into every one of them.
 
-// The handler macro expands a module taking a `params` input into an `extern "C"` wrapper reading
-// the string out of a raw pointer. Clippy reports that on the handler we wrote rather than on the
-// generated wrapper, and the macro drops any attribute we would put on the handler to silence it.
-#![allow(clippy::not_unsafe_ptr_arg_deref)]
-
-mod abi;
-mod params;
+pub mod abi;
+pub mod accounts;
+pub mod params;
+pub mod report;
 
 // Generated by `substreams protogen` out of `proto/`, and committed: regenerating them reaches
 // out to the buf remote plugins, which a plain `cargo build` has no business needing.
@@ -21,590 +25,15 @@ mod params;
 #[allow(clippy::all)]
 pub mod pb;
 
-use params::{BalanceBlock, Params};
-use substreams::errors::Error;
-use substreams::scalar::BigInt;
 use substreams::Hex;
-use substreams_ethereum::pb::eth::rpc::{RpcGetBalanceRequest, RpcGetBalanceRequests};
-use substreams_ethereum::pb::eth::v2 as eth;
-use substreams_ethereum::rpc::RpcBatch;
-use substreams_ethereum::Event;
-
-use pb::sf::ethereum::substreams::ethmethods::v1::{
-    CallTargetSource, Erc20Transfer, Erc721Transfer, EthCallReport, EthGetBalanceReport,
-    EthMethodsReport, InvocationKind, ReportStatus, RpcInvocation, RpcResult, RpcStats, Transfers,
-};
 
 /// `Hex::encode` emits lowercase hexadecimal without a prefix, every address and hash we emit
 /// goes through here so the output is consistently `0x` prefixed.
-fn hex0x(bytes: &[u8]) -> String {
+pub fn hex0x(bytes: &[u8]) -> String {
     format!("0x{}", Hex::encode(bytes))
 }
 
-#[substreams::handlers::map]
-fn map_transfers(block: eth::Block) -> Result<Transfers, Error> {
-    let mut erc20_transfers = Vec::new();
-    let mut erc721_transfers = Vec::new();
-
-    for trx in block.transactions() {
-        let tx_hash = hex0x(&trx.hash);
-
-        for (log, _call) in trx.logs_with_calls() {
-            // The two standards share the same `Transfer` topic and are told apart by their
-            // indexed parameter count, which is exactly what the generated matchers check: the
-            // ERC-20 one wants 3 topics and a 32 bytes payload, the ERC-721 one 4 topics and an
-            // empty payload. The two are mutually exclusive, so the order below is irrelevant.
-            if let Some(transfer) = abi::erc20::events::Transfer::match_and_decode(log) {
-                erc20_transfers.push(Erc20Transfer {
-                    block_number: block.number,
-                    tx_hash: tx_hash.clone(),
-                    log_index: log.index,
-                    contract: hex0x(&log.address),
-                    from: hex0x(&transfer.from),
-                    to: hex0x(&transfer.to),
-                    value: transfer.value.to_string(),
-                });
-                continue;
-            }
-
-            if let Some(transfer) = abi::erc721::events::Transfer::match_and_decode(log) {
-                erc721_transfers.push(Erc721Transfer {
-                    block_number: block.number,
-                    tx_hash: tx_hash.clone(),
-                    log_index: log.index,
-                    contract: hex0x(&log.address),
-                    from: hex0x(&transfer.from),
-                    to: hex0x(&transfer.to),
-                    token_id: transfer.token_id.to_string(),
-                });
-            }
-        }
-    }
-
-    Ok(Transfers {
-        block_number: block.number,
-        erc20_transfers,
-        erc721_transfers,
-    })
-}
-
-#[substreams::handlers::map]
-fn map_eth_call(
-    raw_params: String,
-    block: eth::Block,
-    transfers: Transfers,
-) -> Result<EthCallReport, Error> {
-    let params = Params::parse(&raw_params).map_err(Error::msg)?;
-
-    let rounds = params::rounds_for(params.call_ratio, block.number);
-    if rounds == 0 {
-        return Ok(EthCallReport {
-            block_number: block.number,
-            status: ReportStatus::SkippedByRatio as i32,
-            ..Default::default()
-        });
-    }
-
-    let Some((contract, target_source)) = resolve_call_target(&params, &block, &transfers) else {
-        return Ok(EthCallReport {
-            block_number: block.number,
-            status: ReportStatus::InsufficientBlockData as i32,
-            hints: vec![
-                "this block carries no ERC-20 transfer and no transaction calling a contract, \
-                 so there is nothing to run 'balanceOf' against. Expected on a quiet chain, a \
-                 testnet in particular: widen the range, or set the 'call_to' parameter to a \
-                 contract you know is deployed on this network."
-                    .to_string(),
-            ],
-            ..Default::default()
-        });
-    };
-
-    let mut hints = Vec::new();
-    let accounts = match call_accounts(&contract, &transfers) {
-        Some(accounts) => accounts,
-        None => {
-            // Reached when the target holds no transfer in this block, which is the norm as soon
-            // as `call_to` pins a contract quieter than the chain it sits on. Querying unrelated
-            // accounts still exercises the extension, it just answers zero.
-            hints.push(format!(
-                "{} emitted no transfer in this block, so the accounts queried below are not \
-                 holders and are expected to answer a zero balance. That is a property of the \
-                 block, not a broken endpoint.",
-                hex0x(&contract)
-            ));
-            block_accounts(&block, &transfers)
-        }
-    };
-
-    if accounts.is_empty() {
-        return Ok(EthCallReport {
-            block_number: block.number,
-            status: ReportStatus::InsufficientBlockData as i32,
-            hints: vec![
-                "this block carries no address at all to call 'balanceOf' with. Expected on a \
-                 quiet chain, a testnet in particular: widen the range."
-                    .to_string(),
-            ],
-            target_contract: hex0x(&contract),
-            target_source: target_source as i32,
-            ..Default::default()
-        });
-    }
-
-    if target_source == CallTargetSource::CalledContract {
-        hints.push(format!(
-            "this block carries no ERC-20 transfer, so 'balanceOf' is run against {}, an address \
-             it called with a payload and which therefore holds code. An empty response is the \
-             expected answer when that contract is not a token, and does not indicate a broken \
-             endpoint.",
-            hex0x(&contract)
-        ));
-    }
-
-    let mut cursor = Cursor::new(accounts);
-    let mut invocations = Vec::new();
-
-    for round in 0..rounds {
-        // One single-request invocation and one batched invocation per round: they travel
-        // through the same host function but only the second one exercises the batching of the
-        // underlying JSON-RPC transport.
-        invocations.push(eth_call_invocation(
-            InvocationKind::Single,
-            round,
-            &contract,
-            &cursor.take(1),
-        ));
-
-        if params.call_batch_size > 0 {
-            invocations.push(eth_call_invocation(
-                InvocationKind::Batch,
-                round,
-                &contract,
-                &cursor.take(params.call_batch_size as usize),
-            ));
-        }
-    }
-
-    let stats = stats_of(&invocations);
-    if stats.request_count > 0 && stats.request_count == stats.failure_count {
-        hints.push(format!(
-            "every request failed. Check that the serving instance was started with \
-             --substreams-rpc-endpoints and that the endpoint answers 'eth_call' at {}.",
-            hex0x(&contract)
-        ));
-    } else if invocations
-        .iter()
-        .flat_map(|invocation| &invocation.results)
-        .all(|result| !result.failed && result.value.is_empty())
-    {
-        hints.push(format!(
-            "every request succeeded but answered an empty payload, which is what an address \
-             holding no code returns. {} is most likely not deployed on this network: set \
-             'call_to' to a token that exists here.",
-            hex0x(&contract)
-        ));
-    }
-
-    Ok(EthCallReport {
-        block_number: block.number,
-        status: ReportStatus::Performed as i32,
-        hints,
-        target_contract: hex0x(&contract),
-        target_source: target_source as i32,
-        stats: Some(stats),
-        invocations,
-    })
-}
-
-#[substreams::handlers::map]
-fn map_eth_get_balance(
-    raw_params: String,
-    block: eth::Block,
-    transfers: Transfers,
-) -> Result<EthGetBalanceReport, Error> {
-    let params = Params::parse(&raw_params).map_err(Error::msg)?;
-
-    let rounds = params::rounds_for(params.balance_ratio, block.number);
-    if rounds == 0 {
-        return Ok(EthGetBalanceReport {
-            block_number: block.number,
-            status: ReportStatus::SkippedByRatio as i32,
-            ..Default::default()
-        });
-    }
-
-    let accounts = block_accounts(&block, &transfers);
-    if accounts.is_empty() {
-        return Ok(EthGetBalanceReport {
-            block_number: block.number,
-            status: ReportStatus::InsufficientBlockData as i32,
-            hints: vec![
-                "this block carries no address at all, not even a fee recipient, so there is no \
-                 account to query a balance for. Expected on a quiet chain, a testnet in \
-                 particular: widen the range."
-                    .to_string(),
-            ],
-            ..Default::default()
-        });
-    }
-
-    // Querying the block being processed is the only choice giving a reproducible answer, at the
-    // cost of requiring an archive node when replaying history. `balance_block=latest` trades
-    // that reproducibility for reaching any node.
-    let block_param = match params.balance_block {
-        BalanceBlock::Number => format!("0x{:x}", block.number),
-        BalanceBlock::Latest => "latest".to_string(),
-    };
-
-    let mut cursor = Cursor::new(accounts);
-    let mut invocations = Vec::new();
-
-    for round in 0..rounds {
-        invocations.push(eth_get_balance_invocation(
-            InvocationKind::Single,
-            round,
-            &block_param,
-            &cursor.take(1),
-        ));
-
-        if params.balance_batch_size > 0 {
-            invocations.push(eth_get_balance_invocation(
-                InvocationKind::Batch,
-                round,
-                &block_param,
-                &cursor.take(params.balance_batch_size as usize),
-            ));
-        }
-    }
-
-    let mut hints = Vec::new();
-    let stats = stats_of(&invocations);
-    if stats.request_count > 0 && stats.request_count == stats.failure_count {
-        hints.push(match params.balance_block {
-            // An archive node is what querying an old block needs, and a pruned one answering
-            // nothing else is by far the likeliest reason for a wholesale failure here.
-            BalanceBlock::Number => format!(
-                "every request failed while querying block {block_param}. A node pruning its \
-                 state cannot answer a balance for anything but recent blocks: retry with \
-                 'balance_block=latest', and check that the serving instance was started with \
-                 --substreams-rpc-get-balance-endpoints."
-            ),
-            BalanceBlock::Latest => "every request failed. Check that the serving instance was \
-                 started with --substreams-rpc-get-balance-endpoints and that the endpoint \
-                 answers 'eth_getBalance'."
-                .to_string(),
-        });
-    }
-
-    Ok(EthGetBalanceReport {
-        block_number: block.number,
-        status: ReportStatus::Performed as i32,
-        hints,
-        block_param,
-        stats: Some(stats),
-        invocations,
-    })
-}
-
-#[substreams::handlers::map]
-fn map_eth_methods(
-    eth_call: EthCallReport,
-    eth_get_balance: EthGetBalanceReport,
-) -> Result<EthMethodsReport, Error> {
-    Ok(EthMethodsReport {
-        block_number: eth_call.block_number.max(eth_get_balance.block_number),
-        eth_call: Some(eth_call),
-        eth_get_balance: Some(eth_get_balance),
-    })
-}
-
-/// Rotates over a list of addresses so that successive invocations of a block query different
-/// accounts, wrapping around once the list is exhausted.
-struct Cursor {
-    accounts: Vec<Vec<u8>>,
-    next: usize,
-}
-
-impl Cursor {
-    fn new(accounts: Vec<Vec<u8>>) -> Self {
-        Self { accounts, next: 0 }
-    }
-
-    fn take(&mut self, count: usize) -> Vec<Vec<u8>> {
-        let mut taken = Vec::with_capacity(count);
-        for _ in 0..count {
-            taken.push(self.accounts[self.next].clone());
-            self.next = (self.next + 1) % self.accounts.len();
-        }
-
-        taken
-    }
-}
-
-fn eth_call_invocation(
-    kind: InvocationKind,
-    round: u32,
-    contract: &[u8],
-    accounts: &[Vec<u8>],
-) -> RpcInvocation {
-    let mut batch = RpcBatch::new();
-    for account in accounts {
-        batch = batch.add(
-            abi::erc20::functions::BalanceOf {
-                account: account.clone(),
-            },
-            contract.to_vec(),
-        );
-    }
-
-    let responses = match batch.execute() {
-        Ok(responses) => responses.responses,
-        // The host function itself failing is a setup problem rather than a per-request one,
-        // reporting every request as failed keeps the counters honest.
-        Err(err) => {
-            substreams::log::info!("eth_call invocation failed: {}", err);
-            Vec::new()
-        }
-    };
-
-    let results = accounts
-        .iter()
-        .enumerate()
-        .map(|(index, account)| match responses.get(index) {
-            None => failed_result(account),
-            Some(response) => RpcResult {
-                address: hex0x(account),
-                failed: response.failed,
-                value: RpcBatch::decode::<_, abi::erc20::functions::BalanceOf>(response)
-                    .map(|balance: BigInt| balance.to_string())
-                    .unwrap_or_default(),
-                raw: hex0x(&response.raw),
-            },
-        })
-        .collect();
-
-    invocation_of(kind, round, results)
-}
-
-fn eth_get_balance_invocation(
-    kind: InvocationKind,
-    round: u32,
-    block_param: &str,
-    accounts: &[Vec<u8>],
-) -> RpcInvocation {
-    let requests = RpcGetBalanceRequests {
-        requests: accounts
-            .iter()
-            .map(|account| RpcGetBalanceRequest {
-                address: account.clone(),
-                block: block_param.to_string(),
-            })
-            .collect(),
-    };
-
-    let responses = substreams_ethereum::rpc::eth_get_balance(&requests).responses;
-
-    let results = accounts
-        .iter()
-        .enumerate()
-        .map(|(index, account)| match responses.get(index) {
-            None => failed_result(account),
-            Some(response) => RpcResult {
-                address: hex0x(account),
-                failed: response.failed,
-                value: if response.failed {
-                    String::new()
-                } else {
-                    BigInt::from_unsigned_bytes_be(&response.balance).to_string()
-                },
-                raw: hex0x(&response.balance),
-            },
-        })
-        .collect();
-
-    invocation_of(kind, round, results)
-}
-
-fn failed_result(account: &[u8]) -> RpcResult {
-    RpcResult {
-        address: hex0x(account),
-        failed: true,
-        value: String::new(),
-        raw: String::new(),
-    }
-}
-
-fn invocation_of(kind: InvocationKind, round: u32, results: Vec<RpcResult>) -> RpcInvocation {
-    let failure_count = results.iter().filter(|result| result.failed).count() as u32;
-
-    RpcInvocation {
-        kind: kind as i32,
-        round,
-        request_count: results.len() as u32,
-        success_count: results.len() as u32 - failure_count,
-        failure_count,
-        results,
-    }
-}
-
-fn stats_of(invocations: &[RpcInvocation]) -> RpcStats {
-    RpcStats {
-        invocation_count: invocations.len() as u32,
-        request_count: invocations.iter().map(|it| it.request_count).sum(),
-        success_count: invocations.iter().map(|it| it.success_count).sum(),
-        failure_count: invocations.iter().map(|it| it.failure_count).sum(),
-    }
-}
-
-/// Contract `balanceOf(address)` is called on, and where it was found.
-///
-/// Resolving out of the block, rather than out of a hardcoded token address, is what lets the
-/// package run unchanged against any EVM chain: a pinned mainnet address resolves to an account
-/// holding no code elsewhere, where every call comes back empty and the package reports success
-/// while testing nothing.
-///
-/// The resolution degrades in three steps so that a chain too quiet to hold a token transfer in
-/// every block, which any testnet is, still gets its extension exercised. The step that was
-/// reached is reported so a caller can tell a call into a real token apart from a call into some
-/// contract that merely happened to be there.
-fn resolve_call_target(
-    params: &Params,
-    block: &eth::Block,
-    transfers: &Transfers,
-) -> Option<(Vec<u8>, CallTargetSource)> {
-    if let Some(call_to) = &params.call_to {
-        return Some((call_to.clone(), CallTargetSource::Parameter));
-    }
-
-    if let Some(contract) = busiest_erc20(transfers) {
-        return Some((contract, CallTargetSource::BusiestErc20));
-    }
-
-    // An address a transaction carried a payload to holds code, which is all `balanceOf` needs to
-    // reach the endpoint. It answers empty unless the contract happens to be a token, and that is
-    // still a round trip through the extension, which is what is being tested.
-    block
-        .transactions()
-        .find(|trx| !trx.input.is_empty() && trx.to.len() == 20)
-        .map(|trx| (trx.to.clone(), CallTargetSource::CalledContract))
-}
-
-/// ERC-20 contract that emitted the most `Transfer` events in the block, the accounts seen
-/// transferring it being holders whose balance comes back non-zero.
-fn busiest_erc20(transfers: &Transfers) -> Option<Vec<u8>> {
-    let mut tally: Vec<(&str, usize)> = Vec::new();
-    for transfer in &transfers.erc20_transfers {
-        match tally
-            .iter_mut()
-            .find(|(contract, _)| *contract == transfer.contract)
-        {
-            Some((_, count)) => *count += 1,
-            None => tally.push((transfer.contract.as_str(), 1)),
-        }
-    }
-
-    // Ties are broken on the address itself, leaving the choice independent from the order the
-    // block happens to be iterated in.
-    tally
-        .into_iter()
-        .max_by(|(left_address, left), (right_address, right)| {
-            left.cmp(right).then(right_address.cmp(left_address))
-        })
-        .and_then(|(address, _)| decode_hex_address(address))
-}
-
-/// Holders of the target token seen in this block, whose balance comes back non-zero and so
-/// tells a working call apart from one answering an all-zeroes payload.
-///
-/// `None` when the target moved nothing in this block, leaving the caller to fall back on the
-/// block's own addresses and to say so in its report.
-fn call_accounts(contract: &[u8], transfers: &Transfers) -> Option<Vec<Vec<u8>>> {
-    let contract = hex0x(contract);
-
-    let mut accounts = AddressSet::new();
-    for transfer in &transfers.erc20_transfers {
-        if transfer.contract == contract {
-            accounts.insert_hex(&transfer.from);
-            accounts.insert_hex(&transfer.to);
-        }
-    }
-
-    if accounts.is_empty() {
-        return None;
-    }
-
-    Some(accounts.into_vec())
-}
-
-/// Every distinct address the block touches, senders and recipients of both transactions and
-/// decoded transfers, plus the fee recipient.
-fn block_accounts(block: &eth::Block, transfers: &Transfers) -> Vec<Vec<u8>> {
-    let mut accounts = AddressSet::new();
-
-    for trx in block.transactions() {
-        accounts.insert(&trx.from);
-        accounts.insert(&trx.to);
-    }
-
-    for transfer in &transfers.erc20_transfers {
-        accounts.insert_hex(&transfer.from);
-        accounts.insert_hex(&transfer.to);
-    }
-
-    for transfer in &transfers.erc721_transfers {
-        accounts.insert_hex(&transfer.from);
-        accounts.insert_hex(&transfer.to);
-    }
-
-    // Last, so that a block holding real activity queries that instead. It is however the one
-    // address an empty block still carries, and it is what keeps a chain producing empty blocks,
-    // which any testnet does, from reporting that it has nothing to query.
-    if let Some(header) = &block.header {
-        accounts.insert(&header.coinbase);
-    }
-
-    accounts.into_vec()
-}
-
-/// Deduplicates addresses while preserving the order they were first seen in, so that the
-/// requests a block issues stay identical across runs and across workers.
-struct AddressSet {
-    seen: Vec<Vec<u8>>,
-}
-
-impl AddressSet {
-    fn new() -> Self {
-        Self { seen: Vec::new() }
-    }
-
-    fn insert(&mut self, address: &[u8]) {
-        // Contract creations carry an empty `to`, and the zero address is never worth querying.
-        if address.len() != 20 || address.iter().all(|byte| *byte == 0) {
-            return;
-        }
-
-        if !self.seen.iter().any(|candidate| candidate == address) {
-            self.seen.push(address.to_vec());
-        }
-    }
-
-    fn insert_hex(&mut self, address: &str) {
-        if let Some(address) = decode_hex_address(address) {
-            self.insert(&address);
-        }
-    }
-
-    fn is_empty(&self) -> bool {
-        self.seen.is_empty()
-    }
-
-    fn into_vec(self) -> Vec<Vec<u8>> {
-        self.seen
-    }
-}
-
-fn decode_hex_address(address: &str) -> Option<Vec<u8>> {
+pub fn decode_hex_address(address: &str) -> Option<Vec<u8>> {
     hex::decode(address.strip_prefix("0x").unwrap_or(address))
         .ok()
         .filter(|bytes| bytes.len() == 20)
